pyproject.toml
[project]
name = "macro-cycles"
version = "0.1.0"
description = ""
authors = [
    {name = "zekenewsom",email = "zekenewsom@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "numpy>=1.26,<3.0",
    "fastapi (>=0.118.0,<0.119.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "polars (>=1.34.0,<2.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "duckdb (>=1.4.1,<2.0.0)",
    "pyarrow (>=21.0.0,<22.0.0)",
    "pydantic (>=2.12.0,<3.0.0)",
    "prefect (>=3.4.22,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "tenacity (>=9.1.2,<10.0.0)",
    "pandas-datareader (>=0.10.0,<0.11.0)",
    "yfinance (>=0.2.66,<0.3.0)",
    "fredapi (>=0.5.2,<0.6.0)",
    "hmmlearn (>=0.3.3,<0.4.0)",
    "statsmodels (>=0.14.5,<0.15.0)",
    "scikit-learn (>=1.7.2,<2.0.0)",
    "PyYAML (>=6.0.0,<7.0.0)",
    "pandas-market-calendars (>=5.1.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
# This repository is an application (not a packaged library),
# so disable installing the current project as a package.
package-mode = false

[tool.poetry.scripts]
mc-ingest = "orchestration.flows.ingest_sources:main"
mc-build-indicators = "orchestration.flows.build_indicators:build_indicators"
mc-build-pillars = "orchestration.flows.build_pillars:run"
mc-build-composite = "orchestration.flows.build_composite:run"
mc-build-all = "orchestration.flows.build_pipeline:build_all"
mc-api = "apps.api.run:main"
